refactor(membership): mv join handler to membership module
